+++ biojava-structure/src/test/java/org/biojava/nbio/structure/TestBond.java
@@ -69,7 +69,7 @@ public class TestBond {
 	public void testStructConnModels() throws IOException, StructureException {
 		Structure s = StructureIO.getStructure("1cdr");
 		Group groupOne = s.getPolyChain("A",1).getGroupByPDB(new ResidueNumber("A", 18, ' '));
-		Group groupTwo = s.getNonPolyChain("B",1).getGroupByPDB(new ResidueNumber("B", 1, ' '));
+		Group groupTwo = s.getBranchedChain("B",1).getGroupByPDB(new ResidueNumber("B", 1, ' '));
 		Atom atomOne = groupOne.getAtom("ND2");
 		Atom atomTwo = groupTwo.getAtom("C1");
 		assertTrue(areBonded(atomOne, atomTwo));