+++ biojava-structure/src/test/java/org/biojava/nbio/structure/io/mmtf/TestMmtfRoundTrip.java
@@ -371,14 +371,15 @@ public class TestMmtfRoundTrip {
 		assertEquals(2, structure2.getEntityById(1).getChains().size());
 		assertEquals(2, structure2.getEntityById(2).getChains().size());
 
-		assertEquals(4, structure2.getNonPolyChains().size());
+		assertEquals(4, structure2.getBranchedChains().size());
+		assertEquals(0, structure2.getNonPolyChains().size());
 		assertEquals(4, structure2.getPolyChains().size());
 
 		assertEquals(1, structure2.getEntityById(3).getChains().size());
 
 		// chain asym_id="E" is from entity 3
-		assertSame(structure2.getNonPolyChain("E"), structure2.getEntityById(3).getChains().get(0));
+		assertSame(structure2.getBranchedChain("E"), structure2.getEntityById(3).getChains().get(0));
 
-		assertEquals(5, structure2.getNonPolyChain("E").getAtomGroups().size());
+		assertEquals(5, structure2.getBranchedChain("E").getAtomGroups().size());
 	}
 }